ai-services/rag-service/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP & File Handling
requests==2.31.0
httpx==0.25.0
python-multipart==0.0.6

# Vector Database & Search
chromadb>=0.4.15
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0

# Document Processing
pypdf2>=3.0.1
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.2
pillow>=10.0.0

# Text Processing & Embeddings
sentence-transformers>=2.2.2
tiktoken>=0.5.1
unstructured>=0.10.30

# ML & NLP
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
aiofiles>=23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
